refactor: organiza widgets componentes em pastas
