+++ challenge/rick_morty_app/lib/components/cards/character_card.dart
