+++ challenge/rick_morty_app/lib/components/cards/episode_card.dart
