+++ challenge/rick_morty_app/lib/components/cards/location_card.dart
