+++ challenge/rick_morty_app/lib/components/detailed_cards/detailed_character_card.dart
