+++ challenge/rick_morty_app/lib/components/detailed_cards/detailed_episode_card.dart
