+++ challenge/rick_morty_app/lib/components/filters/filter_character_component.dart
