+++ challenge/rick_morty_app/lib/components/navigation/search_bar_component.dart
