+++ challenge/rick_morty_app/lib/components/navigation/side_bar_component.dart
