+++ challenge/rick_morty_app/lib/pages/locations_page.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
-import 'package:rick_morty_app/components/app_bar_component.dart';
-import 'package:rick_morty_app/components/location_card.dart';
-import 'package:rick_morty_app/components/pagination_bar.dart';
-import 'package:rick_morty_app/components/search_bar_component.dart';
-import 'package:rick_morty_app/components/side_bar_component.dart';
+import 'package:rick_morty_app/components/app_bar/app_bar_component.dart';
+import 'package:rick_morty_app/components/cards/location_card.dart';
+import 'package:rick_morty_app/components/navigation/pagination_bar.dart';
+import 'package:rick_morty_app/components/navigation/search_bar_component.dart';
+import 'package:rick_morty_app/components/navigation/side_bar_component.dart';
 import 'package:rick_morty_app/data/repository.dart';
 import 'package:rick_morty_app/models/paginated_locations.dart';
 import 'package:rick_morty_app/pages/location_details_page.dart';
